Fix dependencies for plotting
Remove HIPS-LIB import in Simple HMM Demo notebook,
and replace it by adding the relevant colormap functions to a new file,
plots.py.

Additionally, add Seaborn to the list of install requirements. This is
to avoid having users manually fix the Seaborn dependency if they try to
run the notebooks.