LDS: Change LDS.sample API to return (x, y)
Because LDS subclasses SLDS, the current way of sampling from an LDS is
to do (_, x, y) = LDS.sample(...). This is a bit strange if the user
doesn't know that internally an LDS is built from an SLDS, and we're
just throwing away the discrete state.

Propose explicitly overriding the sample definition, and returning a
tuple (x, y) instead of a tuple (x, y, z).