Observations: Change API for dynamics of LDS
Previously, setting the dynamics matrix for an LDS required the following:
`lds.dynamics.As[0] = A`. This is because under the hood, we treat LDS as a special case
of SLDS, with just one dynamics matrix.

Propose simplifying the API using properties, so that the dynamics matrix and offset
vector can be set as follows:
`lds.dynamics.dynamics_matrix = A
lds.dynamics.b = b`.