Remove EPS from Logs in ConstrainedTransitions
For the ConstrainedTransitions class, remove the epsilon when taking the log of the transition
matrices. This results in some -infs in the return values of log_transition_matrices, but this
shouldn't interfere with the message passing code. We should do more tests to ensure it also does
not mess up any of the likelihood calculations.